+++ lib/page/game/host/host_page.dart
@@ -78,7 +78,7 @@ class _GamePageState extends ConsumerState<HostPage> {
         body: SafeArea(
           child: Column(
             children: [
-              BannerAdWidget(width: width.toInt(), height: 48),
+              BannerAdWidget(width: width, height: 48),
               Expanded(
                 child: SizedBox(
                   width: width,
